Fixed stats display in event combat rooms
For events like Masked Bandits the logic has to be handled differently
since the room is an EventRoom instead of a MonsterRoom. I updated the
logic and fixed stats display in event combat rooms and on loading into
a completed event combat room. The problem with loading into a completed
event combat room is that the room logic doesn't set the
lastCombatMetricKey to identify the enemy. To fix that I added logic to
save and load FightTracker to the game data, which can conditionally be
used when initializing the FightTracker for the room. This probably
needs more testing around other event combats.