Updating StatsRenderer initial limits to be based on screen height/width
Initialize StatsRenderer in the mod PostInitialize so Settings are
initialized (screen height/width). Attempted to calculate initial stats
Y position based on AbstractRelic positioning so it appears beneath
relics, but I can't seem to get a definitive relic height. For now,
leaving it as a guestimate that should put the top beneath the relic
bar.